Add tests for previewing Repeaters
