+++ packages/studio/tests/__mocks__/entityFile.json
@@ -1,3 +1,4 @@
 {
-  "employeeCount": 123
+  "employeeCount": 123,
+  "favs": ["cat", "dog", "sleep"]
 }
